refactor(app): cleanup
